Added more comments + cleaned up strucutre of MP tests
